move query parser folder
