+++ src/utils/suggester-workers/searching/query-parser/query-parser-soft.js
@@ -1,4 +1,4 @@
-import { prepareStringIndexation } from '../commons-tokenizer';
+import { prepareStringIndexation } from '../../commons-tokenizer';
 
 function parser(string) {
 	return [prepareStringIndexation(string, '-')];
